initial attempt to make libremarkable buildable on other platforms
