+++ Cargo.toml
@@ -43,13 +43,16 @@ hlua = { version = "0.4.1", optional = true }
 stopwatch = { version = "0.0.7", optional = true }
 
 [features]
-default = ["framebuffer", "framebuffer-storage", "framebuffer-drawing", "image", "framebuffer-text-drawing", "input", "battery", "appctx", "hlua"]
+default = ["scan", "framebuffer-types", "framebuffer", "framebuffer-storage", "framebuffer-drawing", "image", "framebuffer-text-drawing", "input-types", "input", "battery", "appctx", "hlua"]
 
-framebuffer = ["memmap2", "ioctl-gen"]
+scan = ["evdev"]
+framebuffer-types = ["ioctl-gen"]
+framebuffer = ["scan", "framebuffer-types", "memmap2"]
 framebuffer-storage = ["framebuffer", "zstd"]
 framebuffer-drawing = ["framebuffer", "line_drawing"]
 framebuffer-text-drawing = ["framebuffer-drawing", "rusttype"]
-input = ["evdev", "epoll", "fxhash"]
+input-types = []
+input = ["scan", "input-types", "evdev", "epoll", "fxhash"]
 battery = []
 appctx = ["framebuffer-text-drawing", "image", "input", "aabb-quadtree"]
