+++ src/framebuffer/refresh.rs
@@ -4,16 +4,10 @@ use std::sync::atomic::Ordering;
 use log::warn;
 
 use crate::framebuffer;
-use crate::framebuffer::common;
+use crate::framebuffer::{common, PartialRefreshMode};
 use crate::framebuffer::core;
 use crate::framebuffer::core::FramebufferUpdate;
 use crate::framebuffer::mxcfb::*;
-
-pub enum PartialRefreshMode {
-    DryRun,
-    Async,
-    Wait,
-}
 
 impl framebuffer::FramebufferRefresh for core::Framebuffer {
     fn full_refresh(
