+++ src/lib.rs
@@ -31,7 +31,7 @@ macro_rules! end_bench {
     };
 }
 
-#[cfg(feature = "framebuffer")]
+#[cfg(feature = "framebuffer-types")]
 #[macro_use(io, ioc, iow, iowr)]
 extern crate ioctl_gen;
 
@@ -49,11 +49,11 @@ pub use line_drawing;
 pub use stopwatch;
 
 /// One of the core components, allowing output and refresh of the EInk display
-#[cfg(feature = "framebuffer")]
+#[cfg(feature = "framebuffer-types")]
 pub mod framebuffer;
 
 /// The other core component, allowing decoding of the three input devices present on the tablet
-#[cfg(feature = "input")]
+#[cfg(feature = "input-types")]
 pub mod input;
 
 /// Device dimensions.
