+++ src/ui_extensions/element.rs
@@ -4,7 +4,7 @@ use std::sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard};
 use crate::framebuffer::cgmath;
 use crate::framebuffer::common;
 use crate::framebuffer::common::{color, mxcfb_rect};
-use crate::framebuffer::refresh::PartialRefreshMode;
+use crate::framebuffer::PartialRefreshMode;
 use crate::framebuffer::FramebufferDraw;
 use crate::framebuffer::FramebufferRefresh;
